test: split Studio object-metadata diagnostics
